Initial commit of the main ipynb file
This is the main program for predicting Boston housing price. 

The dataset for this project originates from the UCI Machine Learning Repository. 

The Boston housing data was collected in 1978 and each of the 506 entries represent aggregated data about 14 features for homes from various suburbs in Boston, Massachusetts. 

This is a prediction problem solved by regression algorithms.